BACKEND_/requirements.txt
fastapi[standard]
prometheus_client
psycopg2-binary
nltk
numpy
onnxruntime
scikit-learn